Moving document generator request to its own model
