+++ src/main/java/uk/gov/companieshouse/api/accounts/model/ixbrl/documentgenerator/DocumentGeneratorRequest.java
@@ -1,10 +1,10 @@
-package uk.gov.companieshouse.api.accounts.model.ixbrl;
+package uk.gov.companieshouse.api.accounts.model.ixbrl.documentgenerator;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.gson.Gson;
 
 
-public class GenerateDocumentRequest {
+public class DocumentGeneratorRequest {
 
     @JsonProperty("resource_uri")
     private String resourceUri;
