+++ src/main/java/uk/gov/companieshouse/api/accounts/model/ixbrl/documentgenerator/Links.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.api.accounts.model.ixbrl;
+package uk.gov.companieshouse.api.accounts.model.ixbrl.documentgenerator;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
